regcharclass: Add tricky fold characters.
The tricky fold characters need to be expanded to include the ones
that map to the same ones as the original set.  This isn't because the
new ones have a length issue, it's that they get left out of comparisons
because of the special regnodes generated for the tricky ones.