make t/op/rand.t fail less often
test 5 grabs many random numbers, and analyses the distribution
of 0 and 1 bits in the returned values.

It currently fails if the standard deviation is > 3.3.
This means it will fail on roughly 0.1% of runs even for a good system.

Given that an individual smoke report may contain 60+ runs, and given the
number of smokers running, this means that we see false negatives
essentially every day.

Increase the SD threshold to 4.0, which means that it will now fail every
0.006% runs; i.e. go from 1 in 1,000 false positives to one in 17,000
(approx).

(Note that just increasing the number of iterations in the run doesn't
change the 0.1% run failure rate.)