Add very simple microtest for microperl
For now, all the tests we run don't use Config, so don't try to build it first.
[Hence this is about half of the original commit by David Leadbeater, which
did build Config, from config.sh]

There's an unresolved problem with building Config - building Config.pm etc
from config.sh won't reflect the capabilities of the microperl being tested,
so that argues for the right solution to be adapting configpm to be able to use
uconfig.sh instead. But doing that means that when building and (re)testing
"regular" perl, we could end up with a build tree with lib/Config.pm "up to
date", as far as make is concerned, but totally the wrong contents in reality.
Tests will fail, heads will be scratched, time will be wasted, etc.