For microperl, add uconfig64.sh, with values suitable for an LP64 platform.
Add instructions on how to use it to README.micro.

There's rather too much size-specific in perl's configuration to make it viable
to have a single uconfig.sh that will work "properly" on both LP32 and LP64
platforms. It seems best to ship the default as LP32 (which is correct for
Win32 (although I don't know if microperl yet builds on Win32)), and provide
just a config.sh file for LP64, as most (all?) 64 bit *nix platforms are LP64,
and can run the shell commands to generate a uconfig.h from it.

(Win64 is LLP64, so uconfig64.sh isn't going to be right for it, so the
inability of Win64 to generate uconfig.h from uconfig64.sh isn't a problem)