Don't copy all of %ENV in prereq.t.
It's sufficient for purposes of enabling prerequisite warnings to
set a localized $ENV{PERL_CORE} to undef rather than making a local
copy of the entire %ENV hash.  This makes the test pass on VMS,
where copying %ENV in toto is not supported.

This fixes a smoke failure, so borrow it from upstream, where it's
already released as part of MM 6.57_10.  6.58 will contain an
additional tweak that works with Perl 5.6, but no need to worry
about that here.