[perl #87708] $tied + $tied and $tied - $tied under ‘use integer’
This is just part of #87708.

This fixes + and - under ‘use integer’ when the same tied scalar is
used for both operands and returns two different values. Before this
commit, get-magic would be called only once and the same value used.
In 5.12.x + just worked but the operands were swapped for -.