PATCH: [perl #87810] BBC Text::MultiMarkdown
A statement should have been outside a block but was inside it.
The indentation was correct, and in a number of times reading
the code I still missed it.

I'm having trouble distilling down the failure scenario into
a simple test case, and am short on tuits right now, so a test
will be committed later.