PATCH: [perl #87812] BBC breaks Pod::Coverage::TrustPod
This patch completes the fixing of this problem.  The problem is that
the failing .t set @INC to exclude lib, and hence couldn't find utf8.pm,
which 5.14 was requiring in places where it previously didn't.  This
patch finishes the job of not requiring utf8.pm in so many places as
were inadvertently added in 5.14.  Commit
3ad98780b4bded02c371c83a668dc8f323e57718 started the job.

This patch changes regcomp.c to not set ANYOF_NONBITMAP_NON_UTF8 where
it inappropriately was.  I don't know what I was thinking when I
originally did what this changes.  In order to match outside the bitmap,
these characters all must match something that requires utf8, such as a
LIGATURE FI.