Give EU::CB 04-base.t an absolute path on VMS.
catfile() with an empty string as a first argument does *not*
produce an absolute path on VMS (and probably only does so by
accident elsewhere).  For purposes of the test, it seems that
any absolute path would do, so this could probably be done
portably, but on the eve of 5.14 let's make a separate code
path for VMS and leave everyone else as is for now.