As part of their switch to a multi-arch library layout, Ubuntu 11.04 (and later, presumably) doesn't keep most libraries (such as -lm) in /lib or /usr/lib.  So we have to ask gcc to tell us where to look.  We don't want gcc's own libraries, however, so we filter those out.
This could be conditional on Ubuntu, but other distributions have
announced their intent follow suit, and this scheme seems to work even
on rather old gcc's.  This unconditionally uses gcc because even if the
user is using another compiler, we still need to find the math library
and friends, and I don't know how other compilers will cope with that
situation.  Still, as an escape hatch, allow Configure command line
overrides to plibpth to bypass this check.