collapse plibpth to one line and remove trailing /
The recent change to hints/linux.sh,
40f026236b9959b7ad3260fedc6c66cd30bb7abc
set the plibpth variable.
It was supposed to set all entries on a single line, but it didn't.
Do it now, and also remove trailing /'s.

(The collapsing is a more robust version of the previous commit,
since reverted, that davem wrote independently).