Make keys $scalar an lvalue
This does a run-time check to see whether $scalar is a hash ref, and
dies if it is not.

This is to keep keys \@_ consistent with keys @_.

I cannot simply use OPf_MOD, since that indicates *potential* lvalue
context (including subroutine args).

So, instead, I take advantage of the fact that OPf_SPECIAL is always
set on the LHS of an assignment (usually to indicate that local()
should not erase the value).