Update the documentation for rand() to note that it's not cryptographically secure due to concerns that end-users are unaware of this and use it in situations where security depends on the strength of the randomness generated.
I'd have been happier getting this patch in earlier in the cycle.
We'd hoped to replace the RNG, but that didn't happen in time, so this
doc update is the "better, still not good" fallback.
(cherry picked from commit 9700c45b7e3bd512c56a8c3607f53e73245de6db)