Remove long-stading limitation from FindBin
The FindBin documentation states as a known bug that it will return
a wrong result if a script of the same name as the current one exists in
$PATH, and is executable.

This patch removes the functionality of searching through $PATH.
According to Graham Barr it was only necessary because the SysV shell on
Sun4OS4 was broken - a system where I can't imagine anybody wanting
(and successfully compiling) a modern perl.

On Linux this part wasn't necessary in the case of `perl -S scriptname',
tests on other platforms are very welcome.

As a side effect, this patch also removes some IO operations, speeding up
FindBin slightly.

This patch is based on discussions with Tina MÃ¼ller.

Further "discussion": http://www.perlmonks.org/?node_id=41213