In S_doparseform(), don't force the pattern to a string. This resolves #79532
Previously S_doparseform() was using SvPV_force(), because the pattern had to
be forced to a string, because the compiled format was stored in the string's
buffer. Now that the compiled format is stored in the magic struct, this isn't
necessary.

Additionally, removing the call to SvPV_force() removes the need to hack with
the SvREADONLY() flag in pp_formline.