"fix bug; also prefix ?? matches with m due to 5.14 deprecation"
I also fixed a bug in the original.  I'm always getting C<eof> vs C<eof()>
swapped in my brain, which is what had happened here.  The old code didn't
do what it said it did because it contrary to the comments didn't reset at
each eof -- because it used the C<eof()> form which is all of ARGV rather
than bare C<eof> for the last file read, and thus each per-file component
of ARGV.

I am concerned about the two paragraphs previous to that, because they
use eof() and I am not perfectly clear that they should.  But I left
them as is.

Jesse asked for "a lot of eyes", so if folks could please look at this
patch and see whether it looks ok, I'd appreciate it.  I did test it
under blead, both with and without the prefixed m on the m?? matches,
which is how I discovered it was buggy with the C<eof()> not C<eof>.

--tom