Make ‘require func()’ work with .pm abs path
As of commit 282b29ee485, pp_requires passes an SV to S_doopen_pm,
instead of char*/length pair.

That commit also used sv_mortalcopy() to copy the sv when trying out a
.pmc extension:
+	SV *const pmcsv = sv_mortalcopy(name);

When the path is absolute, the sv passed to S_doopen_pm is the very sv
that was passed to require. If it was returned from a (non-lvalue)
sub-routine, it will be marked TEMP, so the buffer gets stolen.

After the .pmc file is discovered to be nonexistent, S_doopen_pm then
uses its original sv to open the .pm file. But the buffer has been
stolen, so it’s trying to open undef, which fais.

In the mean time, pp_require still has a pointer to the stolen buffer,
which now has a .pmc extenion, it blithely reports that the .pmc file
cannot be found, not realising that its string has changed out from
under it. (Actually, if the file name were just the right length, it
could be reallocated and we could end up with a crash.)

This patch copies the sv more kindly.